// client/components/TextEditor/types.ts
export type ImageElement = {
  type: "image";
  url: string;
  alt: string;
  children: EmptyText[];
};
// ...
export type ParagraphElement = {
  type: "paragraph";
  align?: string;
  children: Descendant[];
};
// ...
export type EmptyText = {
  text: string;
};